# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Authentication",
        "description": "API de gestion des utilisateurs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/theme/create": {
            "post": {
                "tags": [
                    "Thème"
                ],
                "summary": "Créer un thème",
                "description": "Crée un nouveau thème. Réservé aux administrateurs et employés.",
                "operationId": "f1e7719fac1098803d0ed67b4efc3bec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Noël"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thème créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thème créé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/theme/{id}": {
            "get": {
                "tags": [
                    "Thème"
                ],
                "summary": "Afficher un thème",
                "description": "Retourne les informations d'un thème par son identifiant. Accessible publiquement.",
                "operationId": "65e2037078200c987784017d17b618fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du thème",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thème trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "libelle": {
                                            "type": "string",
                                            "example": "Noël"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Thème introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Thème"
                ],
                "summary": "Modifier un thème",
                "description": "Met à jour le libellé d'un thème existant. Réservé aux administrateurs et employés.",
                "operationId": "850cc2d9d94dcc2bc8789fad54467a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du thème",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Hivernal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thème mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thème mis à jour avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Thème introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Thème"
                ],
                "summary": "Supprimer un thème",
                "description": "Supprime un thème par son identifiant. Réservé aux administrateurs et employés.",
                "operationId": "6072410fc3f095ec8450823c10b3c92c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du thème",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thème supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thème supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "404": {
                        "description": "Thème introuvable"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}